defy libaudio: add disabling of voice path to setMode(NORMAL) case
(binary patch)
Milestone RIL wants to disable the voice path immediately at the end
of a phone call via AudioMgrTAPI AUDMGR_disableVoicePath() call. But if
a call recording session is active, this leads to BP panic on Milestone
 - the voice path needs to be active until the recording has been stopped.
To prevent that, the TAPI call is ignored and the voice path is disabled
only moments later from the setMode(NORMAL) call, which happens while
the call recording can be already stopped.

am_aipcm_voice_call_enable(0) call is added to AudioHardwareMot::setMode
NORMAL case by this patch.

This patch mimics the behaviour of XT702 'call recording enabled' libaudio
in this Defy+ libaudio.

Background info: XT702 PRC libaudio enables the call recording functionality
on Milestone, but there's only Froyo build available, so there are audible
beeps when a call is being recorded. As Gingerbread PRC builds remove
the beeps, we have to adapt the Defy libaudio to have the beepless call
recording.
(Yes, I still have no idea how the beeps are actually done in the Froyo
libaudio, so I was not able to disable it by a binary patch of the
Milestone/XT702 native libaudio so far.)

Change-Id: Ief9bc4b1d8fb423bdc67c101870b4b4eb50073d3